Add focus state toggle on non-input elements
Work with the browser's natural focus system. Only the browser
calls focus() to enter the focus state, if the element is in the
focus state and not an input, then on click or tap
call blur() on the element so it goes through the browser's event
system and back in through the blur listener and handler,
so all states match.