Let browser determine click from touch if no active/touchActive prop
If there is no active or touchActive prop, then let the browser
fully manage what it considers to be a click from a touch interaction
without interference from RI. This results in a better match of
when webkit tap highlight color is active to what results in a click.
Note that RI won't call node.click(), so there may be a
delay in the click.